move test, pair error with json inst, data source yaml
