+++ .gitignore
@@ -12,4 +12,3 @@
 coverage_report/
 .coverage
 *.egg-info/
-.*.sw[a-z]
