+++ spec/data_sources/silva_taxonomy.yaml
@@ -1,0 +1,6 @@
+name: silva_taxonomy
+category: taxonomy
+title: SILVA Taxonomy
+home_url: "https://arb-silva.de"
+data_url: "https://arb-silva.de/no_cache/download/archive/"
+
